// frontend/src/bindings/proto/tricked.ts
export interface MetricRow {
    /**
     * @generated from protobuf field: string run_id = 1
     */
    runId: string;
    /**
     * @generated from protobuf field: int32 step = 2
     */
    step: number;
    /**
     * @generated from protobuf field: optional float total_loss = 3
     */
    totalLoss?: number;
    /**
     * @generated from protobuf field: optional float policy_loss = 4
     */
    policyLoss?: number;
    /**
     * @generated from protobuf field: optional float value_loss = 5
     */
    valueLoss?: number;
    /**
     * @generated from protobuf field: optional float reward_loss = 6
     */
    rewardLoss?: number;
    /**
     * @generated from protobuf field: optional float lr = 7
     */
    lr?: number;
    /**
     * @generated from protobuf field: optional float game_score_min = 8
     */
    gameScoreMin?: number;
    /**
     * @generated from protobuf field: optional float game_score_max = 9
     */
    gameScoreMax?: number;
    /**
     * @generated from protobuf field: optional float game_score_med = 10
     */
    gameScoreMed?: number;
    /**
     * @generated from protobuf field: optional float game_score_mean = 11
     */
    gameScoreMean?: number;
    /**
     * @generated from protobuf field: optional float win_rate = 12
     */
    winRate?: number;
    /**
     * @generated from protobuf field: optional float game_lines_cleared = 13
     */
    gameLinesCleared?: number;
    /**
     * @generated from protobuf field: optional float game_count = 14
     */
    gameCount?: number;
    /**
     * @generated from protobuf field: optional float ram_usage_mb = 15
     */
    ramUsageMb?: number;
    /**
     * @generated from protobuf field: optional float gpu_usage_pct = 16
     */
    gpuUsagePct?: number;
    /**
     * @generated from protobuf field: optional float cpu_usage_pct = 17
     */
    cpuUsagePct?: number;
    /**
     * @generated from protobuf field: optional float disk_usage_pct = 18
     */
    diskUsagePct?: number;
    /**
     * @generated from protobuf field: optional float vram_usage_mb = 19
     */
    vramUsageMb?: number;
    /**
     * @generated from protobuf field: optional float mcts_depth_mean = 20
     */
    mctsDepthMean?: number;
    /**
     * @generated from protobuf field: optional float mcts_search_time_mean = 21
     */
    mctsSearchTimeMean?: number;
    /**
     * @generated from protobuf field: optional float elapsed_time = 22
     */
    elapsedTime?: number;
    /**
     * @generated from protobuf field: optional float network_tx_mbps = 23
     */
    networkTxMbps?: number;
    /**
     * @generated from protobuf field: optional float network_rx_mbps = 24
     */
    networkRxMbps?: number;
    /**
     * @generated from protobuf field: optional float disk_read_mbps = 25
     */
    diskReadMbps?: number;
    /**
     * @generated from protobuf field: optional float disk_write_mbps = 26
     */
    diskWriteMbps?: number;
    /**
     * @generated from protobuf field: optional float policy_entropy = 27
     */
    policyEntropy?: number;
    /**
     * @generated from protobuf field: optional float gradient_norm = 28
     */
    gradientNorm?: number;
    /**
     * @generated from protobuf field: optional float representation_drift = 29
     */
    representationDrift?: number;
    /**
     * @generated from protobuf field: optional float mean_td_error = 30
     */
    meanTdError?: number;
    /**
     * @generated from protobuf field: optional float queue_saturation_ratio = 31
     */
    queueSaturationRatio?: number;
    /**
     * @generated from protobuf field: optional float sps_vs_tps = 32
     */
    spsVsTps?: number;
    /**
     * @generated from protobuf field: optional float queue_latency_us = 33
     */
    queueLatencyUs?: number;
    /**
     * @generated from protobuf field: optional float sumtree_contention_us = 34
     */
    sumtreeContentionUs?: number;
    /**
     * @generated from protobuf field: optional float action_space_entropy = 35
     */
    actionSpaceEntropy?: number;
    /**
     * @generated from protobuf field: optional string layer_gradient_norms = 36
     */
    layerGradientNorms?: string;
    /**
     * @generated from protobuf field: repeated float spatial_heatmap = 37
     */
    spatialHeatmap: number[];
    /**
     * @generated from protobuf field: optional float difficulty = 38
     */
    difficulty?: number;
}
// ...
export interface MetricHistory {
    /**
     * @generated from protobuf field: repeated tricked.MetricRow metrics = 1
     */
    metrics: MetricRow[];
}
// ...
export const MetricRow = new MetricRow$Type();
// ...
class MetricHistory$Type extends MessageType<MetricHistory> {
    constructor() {
        super("tricked.MetricHistory", [
            { no: 1, name: "metrics", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => MetricRow }
        ]);
    }
    create(value?: PartialMessage<MetricHistory>): MetricHistory {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.metrics = [];
        if (value !== undefined)
            reflectionMergePartial<MetricHistory>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: MetricHistory): MetricHistory {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* repeated tricked.MetricRow metrics */ 1:
                    message.metrics.push(MetricRow.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: MetricHistory, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* repeated tricked.MetricRow metrics = 1; */
        for (let i = 0; i < message.metrics.length; i++)
            MetricRow.internalBinaryWrite(message.metrics[i], writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message tricked.MetricHistory
 */
export const MetricHistory = new MetricHistory$Type();